+++ ggc-core/src/ggc/New.java
@@ -5,7 +5,7 @@ public class New extends Notification{
   /** Serial number for serialization. */
   private static final long serialVersionUID = 202110262229L;
 
-  public New(String productKey, int productPrice) {
+  public New(String productKey, double productPrice) {
     super(productKey, productPrice);
   }
